+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
-## 1.0.0
+## 0.1.0
   * Tap refactoring
-  * singer-python version upgrade to 6.1.1
   * Fixes dependabot issues
+  * Updated dependencies to latest versions: requests==2.33.1, singer-python==6.8.0, pylint==4.0.5
+  * Adds `parent-tap-stream-id` as discoverable metadata
 
 
 ## 0.0.1